Add capability to change system clock to work with event time and system current time in the runtime
