+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/config/SiddhiContext.java
@@ -168,4 +168,5 @@ public class SiddhiContext {
     public void setRecordTableHandlerManager(RecordTableHandlerManager recordTableHandlerManager) {
         this.recordTableHandlerManager = recordTableHandlerManager;
     }
+
 }
