+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/util/parser/SiddhiAppParser.java
@@ -30,8 +30,7 @@ import org.wso2.siddhi.core.util.SiddhiAppRuntimeBuilder;
 import org.wso2.siddhi.core.util.SiddhiConstants;
 import org.wso2.siddhi.core.util.ThreadBarrier;
 import org.wso2.siddhi.core.util.snapshot.SnapshotService;
-import org.wso2.siddhi.core.util.timestamp.EventTimeBasedMillisTimestampGenerator;
-import org.wso2.siddhi.core.util.timestamp.SystemCurrentTimeMillisTimestampGenerator;
+import org.wso2.siddhi.core.util.timestamp.TimestampGenerator;
 import org.wso2.siddhi.core.window.Window;
 import org.wso2.siddhi.query.api.SiddhiApp;
 import org.wso2.siddhi.query.api.annotation.Annotation;
@@ -167,8 +166,7 @@ public class SiddhiAppParser {
             if (annotation != null) {
                 String idleTime = null;
                 String increment = null;
-                EventTimeBasedMillisTimestampGenerator timestampGenerator = new
-                        EventTimeBasedMillisTimestampGenerator(siddhiAppContext.getScheduledExecutorService());
+                TimestampGenerator timestampGenerator = new TimestampGenerator(siddhiAppContext);
                 // Get the optional elements of playback annotation
                 for (Element e : annotation.getElements()) {
                     if (SiddhiConstants.ANNOTATION_ELEMENT_IDLE_TIME.equalsIgnoreCase(e.getKey())) {
@@ -208,7 +206,7 @@ public class SiddhiAppParser {
                 siddhiAppContext.setTimestampGenerator(timestampGenerator);
                 siddhiAppContext.setPlayback(true);
             } else {
-                siddhiAppContext.setTimestampGenerator(new SystemCurrentTimeMillisTimestampGenerator());
+                siddhiAppContext.setTimestampGenerator(new TimestampGenerator(siddhiAppContext));
             }
             siddhiAppContext.setSnapshotService(new SnapshotService(siddhiAppContext));
             siddhiAppContext.setElementIdGenerator(new ElementIdGenerator(siddhiAppContext.getName()));
